requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Interactive Brokers
ib_insync>=0.9.86

# Data & Analysis
yfinance>=0.2.18
statsmodels>=0.14.0
scikit-learn>=1.3.0
ta>=0.10.2

# Visualization & Dashboard
matplotlib>=3.7.0
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Utilities
pyyaml>=6.0
loguru>=0.7.0
click>=8.1.0
schedule>=1.2.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Notebooks
jupyter>=1.0.0
